fix(quality): resolve SpotBugs violations + NVD API key (#321)
* fix(quality): resolve SpotBugs violations and add NVD API key config

Fix actionable SpotBugs findings across all modules:
- DLS_DEAD_LOCAL_STORE: remove unused assignment in CreateSellerTerminalCommandHandler
- BX_UNBOXING_IMMEDIATELY_REBOXED: fix Integer autoboxing round-trips in
  PageModelTemplateMapper, PageModelTemplateAdminService, PromotionRuleWriteSupport
- RV_RETURN_VALUE_IGNORED_NO_SIDE_EFFECT: remove no-op classify() call
- SF_SWITCH_NO_DEFAULT: add default case in NotificationPolicy
- NP_GUARANTEED_DEREF / NP_NULL_PARAM_DEREF: add null guards in
  UpsertPageModelCommandHandler, PageModelMapper, TenantPageModelController

Also configure OWASP dependency-check to use NVD_API_KEY env var.


* fix(quality): resolve remaining 4 SpotBugs violations in tchalanet-common

- RCN_REDUNDANT_NULLCHECK_OF_NONNULL_VALUE: remove redundant null check
  on toString() return in GenericTypedIdSerializer
- ST_WRITE_TO_STATIC_FROM_INSTANCE_METHOD: delegate to static method
  in SpringBeans.setApplicationContext
- MS_MUTABLE_ARRAY: make TypedIdRegistry.ENTRIES private, expose via
  all() method that returns a defensive copy
- NP_NULL_ON_SOME_PATH_FROM_RETURN_VALUE: cache getMethod() result
  before dereferencing in ApiResponseBodyAdvice


* fix(quality): resolve remaining checkstyle violations

- Add checkstyle suppressions for domain naming (lot1_3d), Spring Data
  query methods (findByTicket_Id), and intentional switch fallthrough
- Wire suppressions.xml into checkstyle plugin config
- Expand star imports in tchalanet-features (7 files)


---------

+++ tchalanet-server/tchalanet-core/src/main/java/com/tchalanet/server/core/promotion/internal/infra/persistence/adapter/PromotionRuleWriteSupport.java
@@ -249,7 +249,7 @@ class PromotionRuleWriteSupport {
               : tiers.stream()
                   .mapToInt(t -> positiveInt(t, "quantity"))
                   .max()
-                  .orElse(effect.getQuantity()));
+                  .orElse(effect.getQuantity().intValue()));
       effect.setQuantityTiers(tiers);
       return;
     }
@@ -259,7 +259,7 @@ class PromotionRuleWriteSupport {
         params.containsKey("quantityPerStep") ? positiveInt(params, "quantityPerStep") : 1);
     effect.setMaxQuantity(
         params.containsKey("maxQuantity")
-            ? positiveInt(params, "maxQuantity")
+            ? Integer.valueOf(positiveInt(params, "maxQuantity"))
             : effect.getQuantity());
     effect.setQuantityTiers(List.of());
   }
